Added template, renamed existing
Added a users_detail template (in /templates) and matching JSON sample
data (in /fixtures).

Renamed some of the previous templates to bring them in line with the
naming conventions of Django's generic views. Templates are now either
foos_list or foos_detail, to match urls of /foos and /foos/:id.